Animals avoid obstacles by updating their pattern
An animal will now avoid an abstacle if it can't complete its pattern
by picking a new direction. The new direction will affect that portion
of its pattern but keep the rest the same. The new direction will also
be mirrored on the way back to keep the pattern in a loop.
Mimic is now fixed to properly handle a pattern executed in the opposite
direction. Turns out that to execute a pattern in the opposite you can't
reverse the list of moves, instead you need to shift the pattern halfway
through.
Also made a test level all the way at the end. Game is currently set to
start there.